Moved center-on-primary-display to application (global) helpers
